requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
distlib==0.3.9
Django==4.2.16
django-celery-beat==2.7.0
django-debug-toolbar==4.4.5
django-filter==24.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==25.9.2
fakeredis==2.23.3
filelock==3.16.1
frozenlist==1.5.0
gunicorn==22.0.0
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mirakuru==2.5.3
multidict==6.1.0
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
port-for==0.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.8.0
pytest-env==1.1.5
pytest-factoryboy==2.7.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-order==1.3.0
pytest-redis==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sentry-sdk==2.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.5
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.18.3
